add examples for all funcs in sets
