--- sets/example_test.go
@@ -1,115 +1,0 @@
-package sets_test
-
-import (
-	"fmt"
-
-	"github.com/life4/genesis/sets"
-)
-
-func ExampleContains() {
-	s := sets.New(3, 4, 5)
-	result := sets.Contains(s, 4)
-	fmt.Println(result)
-	// Output: true
-}
-
-func ExampleDisjoint() {
-	a := sets.New(3, 4, 5)
-	b := sets.New(5, 6, 7)
-	result := sets.Disjoint(a, b)
-	fmt.Println(result) // a and b both share 5
-	// Output: false
-}
-
-func ExampleDisjointMany() {
-	a := sets.New(3, 4)
-	b := sets.New(5, 6)
-	c := sets.New(6, 7)
-	result := sets.DisjointMany(a, b, c)
-	fmt.Println(result) // b and c both share 6
-	// Output: false
-}
-
-func ExampleEmpty() {
-	s := sets.New[int]()
-	result := sets.Empty(s)
-	fmt.Println(result)
-	// Output: true
-}
-
-func ExampleEqual() {
-	a := sets.New(3, 4, 5)
-	b := sets.New(3, 4, 6)
-	result := sets.Equal(a, b)
-	fmt.Println(result)
-	// Output: false
-}
-
-func ExampleEqualMany() {
-	a := sets.New(3, 4, 5)
-	b := sets.New(3, 4, 5)
-	c := sets.New(3, 4, 6)
-	result := sets.EqualMany(a, b, c)
-	fmt.Println(result)
-	// Output: false
-}
-
-func ExampleIntersect() {
-	a := sets.New(3, 4, 5)
-	b := sets.New(5, 6, 7)
-	result := sets.Intersect(a, b)
-	fmt.Println(result)
-	// Output: true
-}
-
-func ExampleMax() {
-	s := sets.New(3, 6, 4, 5)
-	result, _ := sets.Max(s)
-	fmt.Println(result)
-	// Output: 6
-}
-
-func ExampleMin() {
-	s := sets.New(4, 5, 3, 6)
-	result, _ := sets.Min(s)
-	fmt.Println(result)
-	// Output: 3
-}
-
-func ExampleReduce() {
-	s := sets.New(3, 4, 5)
-	add := func(x, acc int) int { return x + acc }
-	result := sets.Reduce(s, 0, add)
-	fmt.Println(result)
-	// Output: 12
-}
-
-func ExampleSubset() {
-	a := sets.New(4, 5)
-	b := sets.New(3, 4, 5, 6)
-	result := sets.Subset(a, b)
-	fmt.Println(result)
-	// Output: true
-}
-
-func ExampleSum() {
-	s := sets.New(3, 4, 5)
-	result := sets.Sum(s)
-	fmt.Println(result)
-	// Output: 12
-}
-
-func ExampleToSlice() {
-	s := sets.New(3)
-	result := sets.ToSlice(s)
-	fmt.Println(result)
-	// Output: [3]
-}
-
-func ExampleSuperset() {
-	a := sets.New(3, 4, 5, 6)
-	b := sets.New(4, 5)
-	result := sets.Superset(a, b)
-	fmt.Println(result)
-	// Output: true
-}
